add crud for robot shuttle and team add validation for team and captain

+++ Galaxy.Teams.Core/Models/Robot.cs
@@ -8,17 +8,14 @@ namespace Galaxy.Teams.Core.Models
     [Table("Robots")]
     public class Robot: Entity
     {
-        [Column(TypeName = "varchar(256)")]
-        public string Name { get; set; }
+        [Column(TypeName = "varchar(256)")] public string Name { get; set; }
         public RobotStatus Status { get; set; }
-        [Column(TypeName = "varchar(256)")]
-        public string Manufacturer { get; set; }
-        [Column(TypeName = "varchar(256)")]
-        public string Model { get; set; }
+        [Column(TypeName = "varchar(256)")] public string Manufacturer { get; set; }
+        [Column(TypeName = "varchar(256)")] public string Model { get; set; }
         public int Year { get; set; }
-        public float UnitsCoveredInADay { get; set; }
-        public float TrustWorthyPercentage { get; set; }
-        public float FuelConsumptionPerDay { get; set; }
+        public int UnitsCoveredInADay { get; set; }
+        public int TrustWorthyPercentage { get; set; }
+        public int FuelConsumptionPerDay { get; set; }
         public DateTime NextRevision { get; set; }
     }
 }
